feat: add RecordActions component for unified record management
- Introduced a new RecordActions component to handle various actions (view, edit, delete, clone, publish, download) for records in both table and card views.
- Refactored RecordTableView to utilize the new RecordActions component, simplifying the action handling logic.
- Removed redundant action handling code from RecordTableView and moved it to RecordActions for better maintainability.
- Updated index.js to export the new RecordActions component.
