+++ src/components/RecordList/index.js
@@ -1,4 +1,5 @@
 export { default } from './RecordList';
 export { default as RecordList, RecordListProvider, useRecordListContext } from './RecordList';
+export { default as RecordActions } from './RecordActions';
 export { reviewerConfig, publishedConfig, submissionsConfig, createColumns, recordToRow } from './config';
 export { useViewPreference, useColumnVisibility, useResponsiveDefault } from './hooks';
